types: added return type to describe table
